Implements advanced dialog with options

+++ src/weapons/Weapon.cpp
@@ -4,14 +4,14 @@
 void Weapon::useWeapon() {
 
     // set non-usable texture, make usable false and restart the clock
-    weaponSprite.setTextureRect(sf::IntRect({weaponSprite.getTextureRect().position.x - TILE_SIZE, weaponSprite.getTextureRect().position.y}, {TILE_SIZE, TILE_SIZE}));
+    weaponSprite.setTextureRect(sf::IntRect({weaponSprite.getTextureRect().position.x - static_cast<int>(TILE_SIZE), weaponSprite.getTextureRect().position.y}, {TILE_SIZE, TILE_SIZE}));
     usable = false;
     cooldownClock.restart();
 }
 
 void Weapon::update() {
     if (!usable && cooldownClock.getElapsedTime().asMilliseconds() >= cooldown) {
-        weaponSprite.setTextureRect(sf::IntRect({weaponSprite.getTextureRect().position.x + TILE_SIZE, weaponSprite.getTextureRect().position.y}, {TILE_SIZE, TILE_SIZE}));
+        weaponSprite.setTextureRect(sf::IntRect({weaponSprite.getTextureRect().position.x + static_cast<int>(TILE_SIZE), weaponSprite.getTextureRect().position.y}, {TILE_SIZE, TILE_SIZE}));
         usable = true;
     }
 }
